+++ trieste/models/gpflow/__init__.py
@@ -18,7 +18,7 @@ number of :class:`TrainableProbabilisticModel` wrappers for GPflow-based models.
 """
 
 from . import config, optimizer
-from .architectures import build_gpr, build_sgpr, build_svgp, build_vgp_classifier
+from .builders import build_gpr, build_sgpr, build_svgp, build_vgp_classifier
 from .interface import GPflowPredictor
 from .models import GaussianProcessRegression, SparseVariational, VariationalGaussianProcess
 from .sampler import (
